ipset: drop intvl argument from ipset_add
When we try to add an entry with timeout to a set that is not created
with timeout support, the command fails. Unfortunately we cannot figure
out the reason for the failure from the return value of the
ipset_command function, as it will always return -1 on failure.

We could access the error message in the errmsg or report members of the
ipset_session struct, but as it not defined in the header file, we'd
have to redefine it. This would make it hard to maintain compatibility
with different libipset versions.

Since we don't even support changing the value, just drop the intvl
argument entirely. The name is confusing and wrong anyway.

The documentation already recommends to create the set with timeout
support, so for now the timeout value will have to be controlled there.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>