conf: use a single variable in service config
Now that we support multiple topics, we need to be able to specify
multiple topics in the service configuration file as well.

As this file is used for all init systems, we cannot use an array to
specify multiple topics. The easiest way to solve this, is by using a
single variable name for all options. This allows to add as many topics
as you want, and also makes the port option optional.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>