ipset: validate IP addresses before adding them
While experimenting with dropbear to have it report login failures to
MQTT directly, it became clear that the kernel accepts invalid IP
addresses and adds them to the IPset anyway.

Example:
vallumd[2138]: ipset: added 75.74.48 to blacklist4

This is added to the IPset as: 75.74.0.48 timeout 36650

While it should be considered a bug in the tool that publishes this kind
of bad IP addresses to MQTT, accepting them in vallumd would result in
false positives being blacklisted. Validate IP addresses to avoid this.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>