CMake: support pre-releases in version scheme
When increasing the version number to 0.2.0 after branching 0.1, this
would result in silly version numbers like 0.2.0-0.1.1-9-gc88277a.
To avoid this, add another variable with the last tag, and compare it
with the version defined in CMakeLists.txt. If the last tag is less than
the version in CMakeLists.txt, we are on a pre-release of the next
version, so instead of appending the old version tag, append -0_pre.
This way, the package built when the next release is tagged, will
always have a higher version number due to the package release version
being at least 1.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>