mosquitto: add function to parse topic
We currently only support adding an element to an IPset. To support
multiple actions, we could define a message format and parse each
message. Another option is to embed the action in another topic level,
and subscribe to multiple topics using wildcards.

Implement topic parsing to support this. The next topic level defines
the action. Currently only "add" and "del" will be supported. Default
action is "add", to maintain backward compatibility.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>