mosquitto: fix TLS support
The mosquitto library requires a CA file or CA directory to enable TLS.
If neither of these are specified, TLS will silently be disabled. Due to
this behavior, tests with the -T option gave the impression that
enabling TLS worked just fine, while in fact TLS was not enabled at all.

Add an option to specify a CA file to fix this.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>