inet: add function to get address family
This will be used to check if the IP to be added to the set has the same
address family as the set.

Refactor the ip_valid function to use this new function to avoid
duplicating code.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>
(cherry picked from commit a08c861221b50f75fba0cc23351165e3ce073e30)