WIP: nft backend
Initial support for nftables sets using libnftnl. It should support the
same features as ipset, but is considered experimental.

Reason for publishing the work as-is: at least one person showed
interest, and others can now see it's a work in progress.

There is lots TODO:
* code cleanup
* make CI test both backends
* review / improve documentation

Please do *NOT* file issues for this branch; they will be closed. If you
have ideas or suggestions, or want to help, come talk to me on IRC.