CI: add tests using debug container
For the ipset commands to work, and for vallumd to be able to actually
add elements to these ipsets, the vallumd service container must be run
in privileged mode. As this is currently not possible on Codeberg CI,
the ipset commands will fail, and as a result, vallumd will fail early
in the ipset_do() function due to the ipset not existing. Due to this,
we currently cannot test the full code path.

However, these tests can still uncover bugs in parts of the code, so
let's add them like this. They would in fact have uncovered the bugs
fixed in the following commits:

  a83aaaf26b47 ("ipset: fix segfault in exit_error()")
  7276d9241783 ("inet: prevent stack-buffer-overflow in ip_valid()")

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>