init: harden systemd service
Restrict the systemd service as much as possible to harden security.
Done with the help of `systemd-analyze security vallumd.service`.
This will cause a bunch of warnings on systems with older systemd
versions. However, these warnings are non-fatal and can be ignored.
Maintaining a unit file compatible with each of the supported
distributions is too big a burden for an upstream project to carry.
As such, please do not report issues about these warnings.

While at it, order the Service parameters alphabetically.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>