CMake: don't silence pkg_check_modules calls
Previously the pkg_check_modules had been silenced because some distros
had broken pkg-config modules. This was done to prevent build on such
distros, but it wouldn't have failed on those anyway because we're not
using the REQUIRED argument.

Stop silencing those pkg_check_modules calls so that in case a
dependency is actually missing, the user at least gets a hint of what is
going on.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>